pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ashtray"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "George Y. Kussumoto", email = "georgeyk.dev@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click>=8, <9",
  "pyxdg>=0.26",
  "PyGObject>=3.42.2",
]

[project.scripts]
ashtray = "ashtray.cli:cli"

[project.urls]
Documentation = "https://github.com/georgeyk/ashtray"
Issues = "https://github.com/georgeyk/ashtray/issues"
Source = "https://github.com/georgeyk/ashtray"

[tool.hatch.version]
path = "ashtray/__init__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:ashtray tests}"

[tool.coverage.run]
source_pkgs = ["ashtray", "tests"]
branch = true
parallel = true
omit = [
  "ashtray/__init__.py",
]

[tool.coverage.paths]
ashtray = ["ashtray", "*/ashtray/ashtray"]
tests = ["tests", "*/ashtray/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
